update section body to have placeholders for components
