+++ src/components/seda/SedaExplorerHelp.js
@@ -1,0 +1,16 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+
+const SedaExplorerHelp = props => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+SedaExplorerHelp.propTypes = {
+
+}
+
+export default SedaExplorerHelp
