+++ src/views/explorer/index.js
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
 import { loadRouteLocations } from '../../actions/featuresActions';
-import SedaMapChartView from '../../components/seda/SedaMapChartView';
+import SedaExplorer from '../../components/seda/SedaExplorer';
 import SedaPage from '../../components/seda/SedaPage';
 
 const ExplorerView = ({ loadRouteLocations, locations }) => {
@@ -15,7 +15,7 @@ const ExplorerView = ({ loadRouteLocations, locations }) => {
   }, [])
   return (
     <SedaPage>
-      <SedaMapChartView />
+      <SedaExplorer />
     </SedaPage>
   )
 }
